feat: make sidebar responsive
