+++ src/components/Navbar.tsx
@@ -1,3 +1,4 @@
+import { MobileNav } from "@/components/MobileNav";
 import { ThemeToggle } from "@/components/ThemeToggle";
 import { UserMenu } from "@/components/UserMenu";
 import { getCurrentUser } from "@/lib/auth";
@@ -10,9 +11,12 @@ export const Navbar = async () => {
   return (
     <header className="sticky top-0 z-40 w-full border-b border-gray-200 bg-white dark:border-gray-800 dark:bg-gray-900">
       <div className="container flex h-14 items-center">
-        <Link href="/" className="flex items-center justify-start">
-          <Image src="/logo.svg" alt="Logo" width={28} height={28} />
-        </Link>
+        <div className="flex gap-4">
+          <MobileNav />
+          <Link href="/" className="flex items-center justify-start">
+            <Image src="/logo.svg" alt="Logo" width={28} height={28} />
+          </Link>
+        </div>
         <nav className="ml-auto flex items-center gap-4">
           <ThemeToggle />
           <UserMenu user={user} />
